extract config observer
